[이유나] sprint7

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicChannelService.java
@@ -1,28 +1,29 @@
 package com.sprint.mission.discodeit.service.basic;
 
 import com.sprint.mission.discodeit.dto.data.ChannelDto;
-import com.sprint.mission.discodeit.dto.data.UserDto;
 import com.sprint.mission.discodeit.dto.request.PrivateChannelCreateRequest;
 import com.sprint.mission.discodeit.dto.request.PublicChannelCreateRequest;
 import com.sprint.mission.discodeit.dto.request.PublicChannelUpdateRequest;
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.ChannelType;
-import com.sprint.mission.discodeit.entity.Message;
 import com.sprint.mission.discodeit.entity.ReadStatus;
+import com.sprint.mission.discodeit.exception.ChannelException;
+import com.sprint.mission.discodeit.exception.UserException;
 import com.sprint.mission.discodeit.mapper.ChannelMapper;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.MessageRepository;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
-import jakarta.transaction.Transactional;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.time.Instant;
-import java.util.*;
-
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class BasicChannelService implements ChannelService {
@@ -33,40 +34,56 @@ public class BasicChannelService implements ChannelService {
   private final UserRepository userRepository;
   private final ChannelMapper channelMapper;
 
-  @Override
   @Transactional
+  @Override
   public ChannelDto create(PublicChannelCreateRequest request) {
     String name = request.name();
     String description = request.description();
-    Channel channel = new Channel(ChannelType.PUBLIC, name, description);
 
+    log.info("Attempting to create public channel with name: {}, description: {}", name,
+        description);
+
+    Channel channel = new Channel(ChannelType.PUBLIC, name, description);
     channelRepository.save(channel);
+
+    log.info("Public channel with id {} successfully created", channel.getId());
     return channelMapper.toDto(channel);
   }
 
-  @Override
   @Transactional
+  @Override
   public ChannelDto create(PrivateChannelCreateRequest request) {
+    log.info("Attempting to create private channel for participants: {}", request.participantIds());
+
+    if (request.participantIds().isEmpty()) {
+      log.error("Participant list is empty: {}", request.participantIds());
+      throw UserException.invalidUserList(Map.of("participantIds", request.participantIds()));
+    }
+
     Channel channel = new Channel(ChannelType.PRIVATE, null, null);
     channelRepository.save(channel);
 
-    request.participantIds().stream()
-        .map(userId -> userRepository.findById(userId)
-            .map(user -> new ReadStatus(user, channel, channel.getCreatedAt()))
-            .orElseThrow(() -> new RuntimeException("User not found with id: " + userId)))
-        .forEach(readStatusRepository::save);
+    List<ReadStatus> readStatuses = userRepository.findAllById(request.participantIds()).stream()
+        .map(user -> new ReadStatus(user, channel, channel.getCreatedAt()))
+        .toList();
+    readStatusRepository.saveAll(readStatuses);
 
+    log.info("Private channel with id {} successfully created", channel.getId());
     return channelMapper.toDto(channel);
   }
 
+  @Transactional(readOnly = true)
   @Override
   public ChannelDto find(UUID channelId) {
     return channelRepository.findById(channelId)
         .map(channelMapper::toDto)
-        .orElseThrow(
-            () -> new NoSuchElementException("Channel with id " + channelId + " not found"));
+        .orElseThrow(() -> {
+          log.error("Channel with id {} not found", channelId);
+          return ChannelException.channelNotFound(Map.of("channelId", channelId));
+        });
   }
 
+  @Transactional(readOnly = true)
   @Override
   public List<ChannelDto> findAllByUserId(UUID userId) {
     List<UUID> mySubscribedChannelIds = readStatusRepository.findAllByUserId(userId).stream()
@@ -74,44 +91,52 @@ public class BasicChannelService implements ChannelService {
         .map(Channel::getId)
         .toList();
 
-    return channelRepository.findAll().stream()
-        .filter(channel ->
-            channel.getType().equals(ChannelType.PUBLIC)
-                || mySubscribedChannelIds.contains(channel.getId())
-        )
+    return channelRepository.findAllByTypeOrIdIn(ChannelType.PUBLIC, mySubscribedChannelIds)
+        .stream()
         .map(channelMapper::toDto)
         .toList();
   }
 
-  @Override
   @Transactional
+  @Override
   public ChannelDto update(UUID channelId, PublicChannelUpdateRequest request) {
+    log.info("Attempting to update channel with id: {}", channelId);
+
     String newName = request.newName();
     String newDescription = request.newDescription();
 
     Channel channel = channelRepository.findById(channelId)
-        .orElseThrow(
-            () -> new NoSuchElementException("Channel with id " + channelId + " not found"));
+        .orElseThrow(() -> {
+          log.error("Channel with id {} not found", channelId);
+          return ChannelException.channelNotFound(Map.of("channelId", channelId));
+        });
 
     if (channel.getType().equals(ChannelType.PRIVATE)) {
-      throw new IllegalArgumentException("Private channel cannot be updated");
+      log.error("Private channel with id {} cannot be updated", channelId);
+      throw ChannelException.privateChannelUpdateNotAllowed(Map.of("channelId", channelId));
     }
 
     channel.update(newName, newDescription);
 
+    channelRepository.save(channel);
+
+    log.info("Channel with id {} successfully updated", channelId);
     return channelMapper.toDto(channel);
   }
 
-  @Override
   @Transactional
+  @Override
   public void delete(UUID channelId) {
-    Channel channel = channelRepository.findById(channelId)
-        .orElseThrow(
-            () -> new NoSuchElementException("Channel with id " + channelId + " not found"));
+    log.info("Attempting to delete channel with id: {}", channelId);
 
-    messageRepository.deleteAllByChannelId(channel.getId());
-    readStatusRepository.deleteAllByChannelId(channel.getId());
+    if (!channelRepository.existsById(channelId)) {
+      log.error("Channel with id {} not found", channelId);
+      throw ChannelException.channelNotFound(Map.of("channelId", channelId));
+    }
 
+    messageRepository.deleteAllByChannelId(channelId);
+    readStatusRepository.deleteAllByChannelId(channelId);
     channelRepository.deleteById(channelId);
+    log.info("Channel with id {} successfully deleted", channelId);
   }
 }
