+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -4,25 +4,35 @@ import com.sprint.mission.discodeit.dto.data.MessageDto;
 import com.sprint.mission.discodeit.dto.request.BinaryContentCreateRequest;
 import com.sprint.mission.discodeit.dto.request.MessageCreateRequest;
 import com.sprint.mission.discodeit.dto.request.MessageUpdateRequest;
+import com.sprint.mission.discodeit.dto.response.PageResponse;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.entity.Message;
+import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.exception.ChannelException;
+import com.sprint.mission.discodeit.exception.MessageException;
+import com.sprint.mission.discodeit.exception.UserException;
 import com.sprint.mission.discodeit.mapper.MessageMapper;
+import com.sprint.mission.discodeit.mapper.PageResponseMapper;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.MessageRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.MessageService;
 import com.sprint.mission.discodeit.storage.BinaryContentStorage;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
+import java.time.Instant;
 import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.Map;
+import java.util.Optional;
 import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class BasicMessageService implements MessageService {
@@ -30,25 +40,32 @@ public class BasicMessageService implements MessageService {
   private final MessageRepository messageRepository;
   private final ChannelRepository channelRepository;
   private final UserRepository userRepository;
-  private final BinaryContentRepository binaryContentRepository;
   private final MessageMapper messageMapper;
   private final BinaryContentStorage binaryContentStorage;
+  private final BinaryContentRepository binaryContentRepository;
+  private final PageResponseMapper pageResponseMapper;
 
-  @Override
   @Transactional
+  @Override
   public MessageDto create(MessageCreateRequest messageCreateRequest,
       List<BinaryContentCreateRequest> binaryContentCreateRequests) {
-
     UUID channelId = messageCreateRequest.channelId();
     UUID authorId = messageCreateRequest.authorId();
 
+    log.info("Attempting to create message in channel with id: {} by user with id: {}", channelId,
+        authorId);
+
     Channel channel = channelRepository.findById(channelId)
-        .orElseThrow(() -> new NoSuchElementException(
-            "Channel with id " + messageCreateRequest.channelId() + " does not exist"));
+        .orElseThrow(() -> {
+          log.error("Channel with id {} not found", channelId);
+          return ChannelException.channelNotFound(Map.of("channelId", channelId));
+        });
 
     User author = userRepository.findById(authorId)
-        .orElseThrow(() -> new NoSuchElementException(
-            "Author with id " + messageCreateRequest.authorId() + " does not exist"));
+        .orElseThrow(() -> {
+          log.error("Author with id {} not found", authorId);
+          return UserException.userNotFound(Map.of("authorId", authorId));
+        });
 
     List<BinaryContent> attachments = binaryContentCreateRequests.stream()
         .map(attachmentRequest -> {
@@ -59,9 +76,8 @@ public class BasicMessageService implements MessageService {
           BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
               contentType);
           binaryContentRepository.save(binaryContent);
-
           binaryContentStorage.put(binaryContent.getId(), bytes);
-
+          log.info("Successfully saved attachment with file name: {}", fileName);
           return binaryContent;
         })
         .toList();
@@ -70,47 +86,68 @@ public class BasicMessageService implements MessageService {
     Message message = new Message(content, channel, author, attachments);
 
     messageRepository.save(message);
+
+    log.info("Message with id {} successfully created", message.getId());
     return messageMapper.toDto(message);
   }
 
+  @Transactional(readOnly = true)
   @Override
   public MessageDto find(UUID messageId) {
     return messageRepository.findById(messageId)
         .map(messageMapper::toDto)
-        .orElseThrow(
-            () -> new NoSuchElementException("Message with id " + messageId + " not found"));
+        .orElseThrow(() -> {
+          log.error("Message with id {} not found", messageId);
+          return MessageException.messageNotFound(Map.of("messageId", messageId));
+        });
   }
 
+  @Transactional(readOnly = true)
   @Override
-  public List<MessageDto> findAllByChannelId(UUID channelId) {
-    return messageRepository.findAllByChannelId(channelId).stream()
-        .map(messageMapper::toDto)
-        .toList();
+  public PageResponse<MessageDto> findAllByChannelId(UUID channelId, Instant createAt,
+      Pageable pageable) {
+    Slice<MessageDto> slice = messageRepository.findAllByChannelIdWithAuthor(channelId,
+            Optional.ofNullable(createAt).orElse(Instant.now()),
+            pageable)
+        .map(messageMapper::toDto);
+
+    Instant nextCursor = null;
+    if (!slice.getContent().isEmpty()) {
+      nextCursor = slice.getContent().get(slice.getContent().size() - 1).createdAt();
+    }
+
+    return pageResponseMapper.fromSlice(slice, nextCursor);
   }
 
-  @Override
   @Transactional
+  @Override
   public MessageDto update(UUID messageId, MessageUpdateRequest request) {
-    String newContent = request.newContent();
+    log.info("Attempting to update message with id: {}", messageId);
 
+    String newContent = request.newContent();
     Message message = messageRepository.findById(messageId)
-        .orElseThrow(
-            () -> new NoSuchElementException("Message with id " + messageId + " not found"));
+        .orElseThrow(() -> {
+          log.error("Message with id {} not found", messageId);
+          return MessageException.messageNotFound(Map.of("messageId", messageId));
+        });
+
     message.update(newContent);
 
+    log.info("Message with id {} successfully updated", messageId);
     return messageMapper.toDto(message);
   }
 
-  @Override
   @Transactional
+  @Override
   public void delete(UUID messageId) {
-    Message message = messageRepository.findById(messageId)
-        .orElseThrow(
-            () -> new NoSuchElementException("Message with id " + messageId + " not found"));
+    log.info("Attempting to delete message with id: {}", messageId);
 
-    message.getAttachments()
-        .forEach(attachment -> binaryContentRepository.deleteById(attachment.getId()));
+    if (!messageRepository.existsById(messageId)) {
+      log.error("Message with id {} not found", messageId);
+      throw MessageException.messageNotFound(Map.of("messageId", messageId));
+    }
 
     messageRepository.deleteById(messageId);
+    log.info("Successfully deleted message with id: {}", messageId);
   }
 }
