+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -7,71 +7,74 @@ import com.sprint.mission.discodeit.dto.request.UserUpdateRequest;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.entity.UserStatus;
+import com.sprint.mission.discodeit.exception.UserException;
 import com.sprint.mission.discodeit.mapper.UserMapper;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.repository.UserStatusRepository;
 import com.sprint.mission.discodeit.service.UserService;
 import com.sprint.mission.discodeit.storage.BinaryContentStorage;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
 import java.time.Instant;
 import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class BasicUserService implements UserService {
 
   private final UserRepository userRepository;
-  private final BinaryContentRepository binaryContentRepository;
   private final UserStatusRepository userStatusRepository;
   private final UserMapper userMapper;
+  private final BinaryContentRepository binaryContentRepository;
   private final BinaryContentStorage binaryContentStorage;
 
-  @Override
   @Transactional
+  @Override
   public UserDto create(UserCreateRequest userCreateRequest,
       Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
     String username = userCreateRequest.username();
     String email = userCreateRequest.email();
 
+    log.info("Attempting to create user with username: {} and email: {}", username, email);
+
     if (userRepository.existsByEmail(email)) {
-      throw new IllegalArgumentException("User with email " + email + " already exists");
+      log.error("User with email {} already exists", email);
+      throw UserException.emailAlreadyExist(Map.of("email", email));
     }
     if (userRepository.existsByUsername(username)) {
-      throw new IllegalArgumentException("User with username " + username + " already exists");
+      log.error("User with username {} already exists", username);
+      throw UserException.usernameAlreadyExist(Map.of("userName", username));
     }
 
-    BinaryContent nullableProfileId = optionalProfileCreateRequest
+    BinaryContent nullableProfile = optionalProfileCreateRequest
         .map(profileRequest -> {
           String fileName = profileRequest.fileName();
           String contentType = profileRequest.contentType();
           byte[] bytes = profileRequest.bytes();
-
           BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
               contentType);
           binaryContentRepository.save(binaryContent);
-
           binaryContentStorage.put(binaryContent.getId(), bytes);
-
+          log.info("Successfully uploaded profile image for user with username: {}", username);
           return binaryContent;
         })
         .orElse(null);
-
     String password = userCreateRequest.password();
 
-    User user = new User(username, email, password, nullableProfileId);
+    User user = new User(username, email, password, nullableProfile);
     Instant now = Instant.now();
-    UserStatus userStatus = new UserStatus(user, Instant.now());
+    UserStatus userStatus = new UserStatus(user, now);
 
     userRepository.save(user);
-    userStatusRepository.save(userStatus);
 
+    log.info("User with username {} successfully created", username);
     return userMapper.toDto(user);
   }
 
@@ -79,71 +82,77 @@ public class BasicUserService implements UserService {
   public UserDto find(UUID userId) {
     return userRepository.findById(userId)
         .map(userMapper::toDto)
-        .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
+        .orElseThrow(() -> {
+          log.error("User with id {} not found", userId);
+          return UserException.userNotFound(Map.of("userId", userId));
+        });
   }
 
   @Override
   public List<UserDto> findAll() {
-    return userRepository.findAll()
+    return userRepository.findAllWithProfileAndStatus()
         .stream()
         .map(userMapper::toDto)
         .toList();
   }
 
-  @Override
   @Transactional
+  @Override
   public UserDto update(UUID userId, UserUpdateRequest userUpdateRequest,
       Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
+    log.info("Attempting to update user with id: {}", userId);
+
     User user = userRepository.findById(userId)
-        .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
+        .orElseThrow(() -> {
+          log.error("User with id {} not found", userId);
+          return UserException.userNotFound(Map.of("userId", userId));
+        });
 
     String newUsername = userUpdateRequest.newUsername();
     String newEmail = userUpdateRequest.newEmail();
-
     if (userRepository.existsByEmail(newEmail)) {
-      throw new IllegalArgumentException("User with email " + newEmail + " already exists");
+      log.error("User with email {} already exists", newEmail);
+      throw UserException.emailAlreadyExist(Map.of("newEmail", newEmail));
     }
     if (userRepository.existsByUsername(newUsername)) {
-      throw new IllegalArgumentException("User with username " + newUsername + " already exists");
+      log.error("User with username {} already exists", newUsername);
+      throw UserException.usernameAlreadyExist(Map.of("newUsername", newUsername));
     }
 
-    BinaryContent nullableProfileId = optionalProfileCreateRequest
+    BinaryContent nullableProfile = optionalProfileCreateRequest
         .map(profileRequest -> {
-          Optional.ofNullable(user.getProfile())
-              .ifPresent(
-                  existingProfile -> binaryContentRepository.deleteById(existingProfile.getId()));
-
           String fileName = profileRequest.fileName();
           String contentType = profileRequest.contentType();
           byte[] bytes = profileRequest.bytes();
-
           BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
               contentType);
           binaryContentRepository.save(binaryContent);
-
           binaryContentStorage.put(binaryContent.getId(), bytes);
-
+          log.info("Successfully updated profile image for user with username: {}", newUsername);
           return binaryContent;
         })
         .orElse(null);
 
     String newPassword = userUpdateRequest.newPassword();
-    user.update(newUsername, newEmail, newPassword, nullableProfileId);
+    user.update(newUsername, newEmail, newPassword, nullableProfile);
 
+    userRepository.save(user);
+
+    log.info("User with username {} successfully updated", newUsername);
     return userMapper.toDto(user);
   }
 
-  @Override
   @Transactional
+  @Override
   public void delete(UUID userId) {
-    User user = userRepository.findById(userId)
-        .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
+    log.info("Attempting to delete user with id: {}", userId);
 
-    Optional.ofNullable(user.getProfile())
-        .ifPresent(existingProfile -> binaryContentRepository.deleteById(existingProfile.getId()));
-
-    userStatusRepository.deleteByUserId(userId);
+    if (!userRepository.existsById(userId)) {
+      log.error("User with id {} not found", userId);
+      throw UserException.userNotFound(Map.of("userId", userId));
+    }
 
     userRepository.deleteById(userId);
+    log.info("Successfully deleted user with id: {}", userId);
   }
 }
